dri2: fix multi-drawable flip confusion
A flip pending on the fullscreen drawable shouldn't effect a swap on
a windowed drawable.  This shows up if using a GL compositing window
manager (like compiz) compositing a 3d app.  We still need to keep
track of pending flips globally (per display), but use a per-drawable
counter in the decision to queue up.

Signed-off-by: Rob Clark <rob@ti.com>